Refactor render function in app Refactor render method in app and save file structure changes in tests
