+++ lib/components/App.js
@@ -61,77 +61,42 @@ export default class App extends Component {
   }
 
   render() {
-    if (!this.state.city.length) {
-      return (
-        <div className="App">
+    return (
+      <div className="App">
+        <Search 
+          getCity={this.getCity}
+          city={this.state.city}/>
+        { !this.state.city &&
           <div className="welcomeDiv" style={this.state.background}>
             <Welcome />
-            <Search 
-              getCity={this.getCity}
-              city={this.state.city}
-            />
-          </div>
-        </div>
-      );
-    } else if (this.state.error) {
-      return (
-        <div className="App">
-          <div className="gradientDiv" style={this.state.background}>
-            <Search 
-              getCity={this.getCity}
-              city={this.state.city}
-            />
-            <p className="invalidCityWarning">
-              Please enter a valid city! (ex. Denver, CO) 
-            </p>
-            {
-              this.state.currentWeather &&           
-              <CurrentWeather className="current" 
-                weather={this.state.currentWeather}/>
-            }
-          </div>
-          
-          {
-            this.state.sevenHour && 
-            <SevenHourForecast className="sevenHourForecast" 
-              weather={this.state.sevenHour}/>
+          </div>  } 
+        { this.state.city &&
+            <div className="gradientDiv" style={this.state.background}>
+
+              {this.state.error && 
+                <p className="invalidCityWarning">
+                Please enter a valid city! (ex. Denver, CO) 
+                </p>
+              }
+              {
+                this.state.currentWeather &&           
+                <CurrentWeather className="current" 
+                  weather={this.state.currentWeather}/>
+              }
+              {
+                this.state.sevenHour && 
+                <SevenHourForecast className="sevenHourForecast" 
+                  weather={this.state.sevenHour}/>
+              }
+              {
+                this.state.tenDay &&
+                <TenDayForecast className="tenDayForecast" 
+                  weather={this.state.tenDay}/>
+              }
+            </div>
           }
-          
-          {
-            this.state.tenDay &&
-            <TenDayForecast className="tenDayForecast" 
-              weather={this.state.tenDay}/>
-          }
-        </div>
-      );
-    } else {
-      return (
-        <div className="App">
-          <div className="gradientDiv" style={this.state.background}>
-            <Search 
-              getCity={this.getCity}
-              city={this.state.city}
-            />
-            {
-              this.state.currentWeather &&           
-              <CurrentWeather className="current" 
-                weather={this.state.currentWeather}/>
-            }
-          </div>
-          
-          {
-            this.state.sevenHour && 
-            <SevenHourForecast className="sevenHourForecast" 
-              weather={this.state.sevenHour}/>
-          }
-          
-          {
-            this.state.tenDay &&
-            <TenDayForecast className="tenDayForecast" 
-              weather={this.state.tenDay}/>
-          }
-        </div>
-      );
-    }
-  }
+        
+      </div>
+    );
+  } 
 }
