+++ lib/helpers/key.js
@@ -1,0 +1,3 @@
+const key = "d85b8701d1756fa1";
+
+export default key;
